document everything in the core module + remove support of the `From` trait for isize und usize

+++ src/constraint.rs
@@ -186,16 +186,16 @@ where
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Length {
     /// The value must be of an exact length
-    Exact(usize),
+    Exact(u32),
     /// The length of the value must be less than or equal to the specified
     /// maximum
-    Max(usize),
+    Max(u32),
     /// The length of the value must be greater than or equal to the specified
     /// minimum
-    Min(usize),
+    Min(u32),
     /// The length of the value must be between the specified minimum and
     /// maximum (inclusive)
-    MinMax(usize, usize),
+    MinMax(u32, u32),
 }
 
 impl<T> Validate<Length, FieldName> for T
@@ -206,37 +206,42 @@ where
         let length = self.length();
         if let Some((code, expected)) = match *constraint {
             Length::Exact(exact_len) => {
-                if length != exact_len {
+                if length != exact_len as usize {
                     Some((INVALID_LENGTH_EXACT, exact_len))
                 } else {
                     None
                 }
             }
             Length::Max(max) => {
-                if length > max {
+                if length > max as usize {
                     Some((INVALID_LENGTH_MAX, max))
                 } else {
                     None
                 }
             }
             Length::Min(min) => {
-                if length < min {
+                if length < min as usize {
                     Some((INVALID_LENGTH_MIN, min))
                 } else {
                     None
                 }
             }
             Length::MinMax(min, max) => {
-                if length < min {
+                if length < min as usize {
                     Some((INVALID_LENGTH_MIN, min))
-                } else if length > max {
+                } else if length > max as usize {
                     Some((INVALID_LENGTH_MAX, max))
                 } else {
                     None
                 }
             }
         } {
-            Validation::failure(vec![invalid_value(code, name, self.length(), expected)])
+            Validation::failure(vec![invalid_value(
+                code,
+                name,
+                self.length() as u32,
+                expected,
+            )])
         } else {
             Validation::success(self)
         }
@@ -252,16 +257,16 @@ where
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum CharCount {
     /// The number of characters must be equal to the specified amount
-    Exact(usize),
+    Exact(u32),
     /// The number of characters must be less than or equal to the specified
     /// maximum
-    Max(usize),
+    Max(u32),
     /// The number of characters must be greater than or equal to the specified
     /// minimum
-    Min(usize),
+    Min(u32),
     /// The number of characters must be between the specified minimum and
     /// maximum (inclusive)
-    MinMax(usize, usize),
+    MinMax(u32, u32),
 }
 
 impl<T> Validate<CharCount, FieldName> for T
@@ -276,37 +281,42 @@ where
         let char_count = self.char_count();
         if let Some((code, expected)) = match *constraint {
             CharCount::Exact(exact_val) => {
-                if char_count != exact_val {
+                if char_count != exact_val as usize {
                     Some((INVALID_CHAR_COUNT_EXACT, exact_val))
                 } else {
                     None
                 }
             }
             CharCount::Max(max) => {
-                if char_count > max {
+                if char_count > max as usize {
                     Some((INVALID_CHAR_COUNT_MAX, max))
                 } else {
                     None
                 }
             }
             CharCount::Min(min) => {
-                if char_count < min {
+                if char_count < min as usize {
                     Some((INVALID_CHAR_COUNT_MIN, min))
                 } else {
                     None
                 }
             }
             CharCount::MinMax(min, max) => {
-                if char_count < min {
+                if char_count < min as usize {
                     Some((INVALID_LENGTH_MIN, min))
-                } else if char_count > max {
+                } else if char_count > max as usize {
                     Some((INVALID_CHAR_COUNT_MAX, max))
                 } else {
                     None
                 }
             }
         } {
-            Validation::failure(vec![invalid_value(code, name, self.char_count(), expected)])
+            Validation::failure(vec![invalid_value(
+                code,
+                name,
+                self.char_count() as u32,
+                expected,
+            )])
         } else {
             Validation::success(self)
         }
